+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/AccessoryDesired.java
@@ -107,7 +107,7 @@ public class AccessoryDesired extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			AccessoryDesired obj = new AccessoryDesired();
@@ -121,13 +121,13 @@ public class AccessoryDesired extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public AccessoryDesired GetInstance(UAVObjectManager objMngr, int instID)
+	public AccessoryDesired GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (AccessoryDesired)(objMngr.getObject(AccessoryDesired.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0xC409985A;
+	protected static final long OBJID = 0xC409985A;
 	protected static final String NAME = "AccessoryDesired";
 	protected static String DESCRIPTION = "Desired Auxillary actuator settings.  Comes from @ref ManualControlModule.";
 	protected static final boolean ISSINGLEINST = 0 == 1;
