+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/FlightPlanStatus.java
@@ -152,7 +152,7 @@ public class FlightPlanStatus extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			FlightPlanStatus obj = new FlightPlanStatus();
@@ -166,13 +166,13 @@ public class FlightPlanStatus extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public FlightPlanStatus GetInstance(UAVObjectManager objMngr, int instID)
+	public FlightPlanStatus GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (FlightPlanStatus)(objMngr.getObject(FlightPlanStatus.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0x2206EE46;
+	protected static final long OBJID = 0x2206EE46;
 	protected static final String NAME = "FlightPlanStatus";
 	protected static String DESCRIPTION = "Status of the flight plan script";
 	protected static final boolean ISSINGLEINST = 1 == 1;
